Tidy up method locations to prep for shared folder
