+++ client-v2/src/actions/FrontsConfig.js
@@ -8,18 +8,19 @@ import { getCollectionConfig } from 'selectors/frontsSelectors';
 import {
   getCollectionArticles,
   fetchFrontsConfig,
-  type FrontsConfig,
   getCollection
 } from 'services/faciaApi';
+import type { FrontsConfig } from 'types/FaciaApi';
 import { normaliseCollectionWithNestedArticles } from 'util/shared';
+import {
+  combineCollectionWithConfig,
+  populateDraftArticles
+} from 'util/frontsUtils';
 import { articleFragmentsReceived } from 'actions/ArticleFragments';
 import { externalArticlesReceived } from 'actions/ExternalArticles';
 import {
   collectionReceived,
-  requestFrontCollection,
-  combineCollectionWithConfig,
-  errorReceivingFrontCollection,
-  populateDraftArticles
+  errorReceivingFrontCollection
 } from './Collection';
 
 function frontsConfigReceived(config: FrontsConfig): Action {
@@ -46,9 +47,8 @@ function errorReceivingConfig(error: string): Action {
 }
 
 function getFrontCollection(collectionId: string) {
-  return (dispatch: Dispatch, getState: () => State) => {
-    dispatch(requestFrontCollection());
-    return getCollection(collectionId)
+  return (dispatch: Dispatch, getState: () => State) =>
+    getCollection(collectionId)
       .then((res: Object) => {
         const collectionConfig = getCollectionConfig(getState(), collectionId);
         const collectionWithNestedArticles = combineCollectionWithConfig(
@@ -72,6 +72,7 @@ function getFrontCollection(collectionId: string) {
         );
         return collectionWithDraftArticles.live
           ? [
+              // We use a set to dedupe our article ids
               ...new Set([
                 ...collectionWithDraftArticles.draft.map(
                   nestedArticleFragment => nestedArticleFragment.id
@@ -87,7 +88,6 @@ function getFrontCollection(collectionId: string) {
         dispatch(errorReceivingFrontCollection(error));
         return [];
       });
-  };
 }
 
 const getCollectionsAndArticles = (collectionIds: Array<string>) => (
