+++ client-v2/src/util/frontsUtils.js
@@ -1,6 +1,7 @@
 // @flow
 
 import type { FrontConfig, CollectionConfig } from 'services/faciaApi';
+import type { CollectionWithNestedArticles } from 'types/Shared';
 
 const getFrontCollections = (
   frontId: ?string,
@@ -23,4 +24,21 @@ const getFrontCollections = (
   return [];
 };
 
-export { getFrontCollections };
+const combineCollectionWithConfig = (
+  collectionConfig: CollectionConfig,
+  collection: CollectionWithNestedArticles
+): CollectionWithNestedArticles =>
+  Object.assign({}, collection, {
+    id: collection.id,
+    displayName: collectionConfig.displayName,
+    groups: collectionConfig.groups
+  });
+
+const populateDraftArticles = (collection: CollectionWithNestedArticles) =>
+  !collection.draft ? collection.live : collection.draft;
+
+export {
+  getFrontCollections,
+  combineCollectionWithConfig,
+  populateDraftArticles
+};
